Separate player station interaction from navigation component
