+++ Underworld Diner/Assets/Scripts/Gameplay/Player/Signals/DestinationReachedSignal.cs
@@ -1,0 +1,7 @@
+﻿namespace Gameplay.Player.Signals
+{
+    public class DestinationReachedSignal
+    {
+        
+    }
+}
